cloudbuild/scripts/requirements.txt
pyyaml>=6.0
requests>=2.31
pytest>=7.0
beautifulsoup4>=4.12
google-cloud-aiplatform>=1.38
